requirements.txt
Flask==3.0.0
mem0ai>=1.0.0
openai>=1.0.0
gunicorn==21.2.0

